mood_diary/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
pytest==7.4.3
pytest-cov==4.1.0
bcrypt==4.0.1
email-validator==2.1.0.post1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
matplotlib==3.8.0
pandas==2.1.1 